closure-control: move code to new cluster headers while keeping backward
